+++ lib/src/result/capture_transformer.dart
@@ -4,14 +4,13 @@
 
 import 'dart:async';
 
-import '../result.dart';
+import 'result.dart';
 import 'capture_sink.dart';
 
 /// A stream transformer that captures a stream of events into [Result]s.
 ///
 /// The result of the transformation is a stream of [Result] values and no
-/// error events. This is the transformer used by [captureStream].
-@Deprecated("Will be removed in async 2.0.0.")
+/// error events. Exposed by [Result.captureStream].
 class CaptureStreamTransformer<T> implements StreamTransformer<T, Result<T>> {
   const CaptureStreamTransformer();
 
@@ -19,7 +18,6 @@ class CaptureStreamTransformer<T> implements StreamTransformer<T, Result<T>> {
     return new Stream<Result<T>>.eventTransformed(source, _createSink);
   }
 
-  static EventSink _createSink(EventSink<Result> sink) {
-    return new CaptureSink(sink);
-  }
+  // Since Stream.eventTransformed is not generic, this method can be static.
+  static EventSink _createSink(EventSink<Result> sink) => new CaptureSink(sink);
 }