+++ CHANGELOG.md
@@ -1,7 +1,34 @@
-## 1.13.2
+## 2.0.3
 
 * Fix a bug in `StreamQueue.startTransaction()` and related methods when
   rejecting a transaction that isn't the oldest request in the queue.
+
+## 2.0.2
+
+* Add support for Dart 2.0 library changes to class `Timer`. 
+
+## 2.0.1
+
+* Fix a fuzzy arrow type warning.
+
+## 2.0.0
+* Remove deprecated public `result.dart` and `stream_zip.dart` libraries and
+  deprecated classes `ReleaseStreamTransformer` and `CaptureStreamTransformer`.
+
+* Add `captureAll` and `flattenList` static methods to `Result`.
+
+* Change `ErrorResult` to not be generic and always be a `Result<Null>`.
+  That makes an error independent of the type of result it occurs instead of.
+
+## 1.13.3
+
+* Make `TypeSafeStream` extend `Stream` instead of implementing it. This ensures
+  that new methods on `Stream` are automatically picked up, they will go through
+  the `listen` method which type-checks every event.
+  
+## 1.13.2
+
+* Fix a type-warning.
 
 ## 1.13.1